+++ src/main/java/com/musala/dronemanagement/repository/MedicationRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @Author Geeth_G
  * @CreatedTime on 22-10-2022 1:57:38 PM
  */
-public interface MedicationRepository extends JpaRepository<Medication, Long> {
+public interface MedicationRepository extends JpaRepository<Medication, Integer> {
 }
